+++ src/games/utils/Animated/AnimatedContainer.hpp
@@ -17,7 +17,7 @@ namespace arc {
 	class AnimatedContainer : public AnimatedObject {
 	public:
 		AnimatedContainer(const std::vector<std::SPTR<IObject>> &obj,
-				size_t update);
+			float update);
 
 		std::SPTR<IShape>
 		draw(const std::SPTR<arc::IShape> &parent) const override;
